refact: actions 로직 분리 및 기타 코드 수정
