+++ src/app/(auth)/layout.tsx
@@ -1,6 +1,6 @@
 import { Inter as FontSans } from "next/font/google";
 import { cn } from "@/lib/utils";
-import "../../styles/elisa.css";
+import "../../styles/globals.css";
 import { buttonVariants } from "@/components/shadcn-ui/button";
 import Link from "next/link";
